Converted PAYHeaderView to autolayout

+++ PAYFormBuilder/view/PAYHeaderView.h
@@ -45,8 +45,8 @@
 @property (nonatomic, retain) PAYTextLabel *subTitleLabel;
 
 
-@property (nonatomic, assign) NSUInteger iconTopMargin      UI_APPEARANCE_SELECTOR;
-@property (nonatomic, assign) NSUInteger titleTopMargin     UI_APPEARANCE_SELECTOR;
-@property (nonatomic, assign) NSUInteger subTitleTopMargin  UI_APPEARANCE_SELECTOR;
+@property (nonatomic, assign) CGFloat iconTopMargin      UI_APPEARANCE_SELECTOR;
+@property (nonatomic, assign) CGFloat titleTopMargin     UI_APPEARANCE_SELECTOR;
+@property (nonatomic, assign) CGFloat subTitleTopMargin  UI_APPEARANCE_SELECTOR;
 
 @end
